history now works + simd accessing memory outside of allocation bug

+++ actual_code/History.h
@@ -114,8 +114,7 @@ enum HistoryDirection
 	history_dir_left,
 	history_dir_right,
 };
-
-void log_move(History *history, int8_t direction, int cursor_id, int view_id);
+void log_move(History *history, int direction, int cursor_id, int view_id);
 void log_add(History *history, int8_t direction, char byte, int cursor_id, int view_id);
 void log_remove(History *history, int direction, char byte, int cursor_id, int view_id);
 void log_add_cursor(History *history, int pos, bool is_selection, int cursor_id, int view_id);
